You can now move a character and see it move on multiple clients.
There are few issues, I need it so that that Unit makes the commands and
adds them to the player controller as to make it possible for Units to
have special ablities. I also need to work out away so that it is clear
what team you are and know what team each unit is. I don't know how I am
going to do this.